BodySLAM_Refactored/requirements.txt
pillow==10.4.0
timm==0.6.7
matplotlib